[WIP] - Add new researchers form
